refactor(cli): rename execute module to command
